декомпозиция FilmChooseSeat
