+++ src/pages/FilmDetailPage/components/film-choose-seat/FilmChooseSeat.tsx
@@ -1,59 +1,26 @@
-import { FC, useEffect, useMemo, useState } from 'react';
+import { FC, useEffect } from 'react';
 
 import st from './FilmChooseSeat.module.scss';
 import { ScheduleState } from '../../types/ScheduleState';
 import { useGetSсheduleByidQuery } from '@/services/film-sevice';
 import { useParams } from 'react-router-dom';
 import { Seats } from '@/components/Seats';
-import { IChoosePlace } from '@/shared/types';
+import { useCurrentTime } from './hooks/useCurrentTime';
+import { useChoose } from './hooks/useChoose';
 
 interface FilmChooseSeatProps {
   schedule: ScheduleState;
 }
 
 const FilmChooseSeat: FC<FilmChooseSeatProps> = ({ schedule }) => {
-  const [chooseSeats, setChooseSeats] = useState<IChoosePlace[]>([]);
-
   const params = useParams<{ id: string }>();
   const { data } = useGetSсheduleByidQuery(params?.id || '');
 
-  const currentSeanse = useMemo(() => {
-    if (data)
-      return data.schedules.find((seanse) => seanse.date === schedule.date);
-    return undefined;
-  }, [data, schedule.date]);
-
-  const currentHall = useMemo(() => {
-    if (currentSeanse)
-      return currentSeanse.seances.filter(
-        (seanse) => seanse.hall.name === schedule.hall,
-      );
-    return undefined;
-  }, [currentSeanse, schedule.hall]);
-
-  const currentTime = useMemo(() => {
-    if (currentHall)
-      return currentHall.find((hall) => hall.time === schedule.time);
-    return undefined;
-  }, [currentHall, schedule.time]);
-
-  const onSeatClick = (seat: IChoosePlace) => {
-    const currentSeat = chooseSeats.find(
-      (arrSeat) => arrSeat.row === seat.row && arrSeat.column === seat.column,
-    );
-
-    if (currentSeat)
-      setChooseSeats(
-        chooseSeats.filter(
-          (arrSeat) =>
-            arrSeat.row !== seat.row || arrSeat.column !== seat.column,
-        ),
-      );
-    else setChooseSeats([...chooseSeats, seat]);
-  };
+  const currentTime = useCurrentTime(data?.schedules, schedule);
+  const { chooseSeats, onSeatClick, clearChooseSeats } = useChoose();
 
   useEffect(() => {
-    setChooseSeats([]);
+    clearChooseSeats();
   }, [schedule]);
 
   return (
